pyproject.toml
[project]
name = "channels-backend"
version = "0.1.0"
description = "Python WebSocket server for Discord-like real-time messaging"
authors = [
    {name = "Shalev Asor", email = "shalevasor@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.5",
]

# Tool configurations
[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true